compilation fix for 64 bit systems
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@1997 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/rsiface/rstypes.h
@@ -300,7 +300,7 @@ class CompressedChunkMap
 		uint32_t filledChunks(uint32_t nbchks)
 		{
 			uint32_t res = 0 ;
-			for(uint32_t i=0;i<std::min(nbchks,_map.size()*32);++i)
+			for(uint32_t i=0;i<std::min(nbchks,(uint32_t)_map.size()*32);++i)
 				res += operator[](i) ;
 			return res ;
 		}
